requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
aiosqlite==0.19.0
riotwatcher==3.3.1 